2.Convolutional Neural Network/requirements.txt
numpy==1.22.1
wandb==0.12.10
tqdm
tensorflow-gpu
tensorflow-addons
matplotlib